fix: lock can be released even when not holding the lock
