+++ packages/shared/lib/index.ts
@@ -33,6 +33,7 @@ export * from './services/invitations.js';
 export * from './services/providers.js';
 
 export * as oauth2Client from './clients/oauth2.client.js';
+export * from './clients/locking.js';
 
 export * from './services/nango-config.service.js';
